test(cron): import onTimer from timer test-support after module split
Upstream moved onTimer and executeJobCore out of timer.ts into the
timer.test-support.ts helper module, so the leftover imports from
timer.js broke tsgo type checking (TS2459). Point the three affected
cron test files at ./timer.test-support.js to match the new layout.
